pt_068_imdbpy_test_20210911/requirements.txt
greenlet==1.1.1
IMDbPY==2021.4.18
lxml==4.6.3
SQLAlchemy==1.4.23
